feat: upgrade admin dashboard with modern UI and remove new-dashboard
🚀 Major admin dashboard improvements:

✨ Features:
- Remove deprecated 'Dashboard Mới' page and navigation
- Upgrade main dashboard (/admin) with modern shadcn/ui components
- Integrate StatCard, SystemHealth, ActivityFeed, and QuickActions
- Add responsive design for all screen sizes (mobile, tablet, desktop)
- Implement accessibility features (ARIA labels, keyboard navigation)

🎨 UI/UX Improvements:
- Modern gradient header consistent with other admin pages
- Enhanced statistics cards with growth indicators
- Real-time system health monitoring
- Interactive activity feed with user avatars
- Quick action shortcuts for admin functions

⚡ Performance:
- Optimize with React.memo, useMemo, and useCallback
- Improve loading states and error handling
- Clean TypeScript interfaces and proper state management

🔧 Technical:
- Remove unused new-dashboard route and files
- Update navigation in admin layout
- Maintain ESLint compliance and TypeScript strict mode
- Add comprehensive documentation

📱 Responsive Design:
- Mobile-first approach with proper breakpoints
- Consistent spacing and layout across devices
- Optimized for accessibility and screen readers

Files changed:
- Modified: frontend/src/app/admin/page.tsx (complete rewrite)
- Modified: frontend/src/app/admin/layout.tsx (remove new-dashboard nav)
- Deleted: frontend/src/app/admin/new-dashboard/page.tsx
- Deleted: frontend/docs/NEW_DASHBOARD_GUIDE.md
- Added: frontend/docs/DASHBOARD_UPGRADE_SUMMARY.md

✅ Build tested and verified successful
