--- frontend/docs/NEW_DASHBOARD_GUIDE.md
@@ -1,327 +1,0 @@
-# 🚀 New Admin Dashboard - Complete Guide
-
-## 📋 **OVERVIEW**
-
-Trang dashboard admin mới được xây dựng hoàn toàn từ đầu với shadcn/ui, cung cấp trải nghiệm hiện đại, responsive và accessible cho admin users.
-
-**URL**: `/admin/new-dashboard`
-
----
-
-## 🎯 **KEY FEATURES**
-
-### ✨ **Modern UI Components**
-- **StatCard**: Thẻ thống kê với animations và hover effects
-- **SystemHealth**: Monitoring hệ thống real-time
-- **ActivityFeed**: Feed hoạt động với user avatars
-- **QuickActions**: Shortcuts đến các chức năng chính
-
-### 📱 **Responsive Design**
-- **Mobile (320px+)**: Single column layout
-- **Tablet (768px+)**: Two column grid
-- **Desktop (1024px+)**: Multi-column layout với sidebar
-- **Large screens (1440px+)**: Optimized spacing
-
-### ♿ **Accessibility Features**
-- ARIA labels cho screen readers
-- Keyboard navigation support
-- High contrast mode support
-- Focus indicators
-- Reduced motion support
-
-### ⚡ **Performance Optimizations**
-- React.memo cho components
-- useMemo cho expensive calculations
-- Lazy loading cho data
-- Optimized bundle size (9.12kB)
-
----
-
-## 🏗️ **ARCHITECTURE**
-
-### **Component Structure**
-```
-/src/app/admin/new-dashboard/
-├── page.tsx                 # Main dashboard page
-└── /src/components/admin/dashboard/
-    ├── StatCard.tsx         # Statistics cards
-    ├── SystemHealth.tsx     # System monitoring
-    ├── ActivityFeed.tsx     # Activity timeline
-    └── QuickActions.tsx     # Quick action buttons
-```
-
-### **Data Flow**
-1. **Initial Load**: Fetch all dashboard data
-2. **Auto Refresh**: Every 5 minutes
-3. **Manual Refresh**: User-triggered
-4. **Real-time Updates**: WebSocket (future)
-
----
-
-## 🧪 **TESTING GUIDE**
-
-### **1. Responsive Testing**
-
-#### **Mobile Testing (320px - 768px)**
-```bash
-# Test breakpoints
-- 320px: iPhone SE
-- 375px: iPhone 12
-- 414px: iPhone 12 Pro Max
-- 768px: iPad
-```
-
-**Expected Behavior:**
-- ✅ Single column layout
-- ✅ Collapsible navigation
-- ✅ Touch-friendly buttons (44px min)
-- ✅ Readable text (16px min)
-- ✅ Proper spacing
-
-#### **Tablet Testing (768px - 1024px)**
-```bash
-# Test breakpoints
-- 768px: iPad Portrait
-- 1024px: iPad Landscape
-```
-
-**Expected Behavior:**
-- ✅ Two column grid for stats
-- ✅ Adaptive content layout
-- ✅ Proper touch targets
-
-#### **Desktop Testing (1024px+)**
-```bash
-# Test breakpoints
-- 1024px: Small laptop
-- 1440px: Standard desktop
-- 1920px: Large desktop
-```
-
-**Expected Behavior:**
-- ✅ Four column stats grid
-- ✅ Three column main layout
-- ✅ Hover effects working
-- ✅ Keyboard navigation
-
-### **2. Functionality Testing**
-
-#### **Data Loading**
-```typescript
-// Test scenarios
-1. Initial page load
-2. Refresh button click
-3. Auto-refresh (5 minutes)
-4. Error handling
-5. Loading states
-```
-
-#### **Interactive Elements**
-```typescript
-// Test all clickable elements
-1. Stat cards → Should show details/navigate
-2. Quick actions → Should navigate to correct pages
-3. Activity items → Should show user info
-4. System health → Should show service details
-5. Refresh button → Should reload data
-```
-
-#### **Accessibility Testing**
-```bash
-# Keyboard navigation
-- Tab through all interactive elements
-- Enter/Space to activate buttons
-- Escape to close modals/dropdowns
-
-# Screen reader testing
-- Use NVDA/JAWS/VoiceOver
-- Check ARIA labels
-- Verify heading structure
-```
-
-### **3. Performance Testing**
-
-#### **Bundle Size**
-```bash
-npm run build
-# Check: /admin/new-dashboard should be ~9.12kB
-```
-
-#### **Loading Performance**
-```bash
-# Lighthouse audit
-npm run dev
-# Navigate to /admin/new-dashboard
-# Run Lighthouse performance audit
-# Target: 90+ score
-```
-
-#### **Memory Usage**
-```bash
-# Chrome DevTools
-1. Open Performance tab
-2. Record page interaction
-3. Check for memory leaks
-4. Verify smooth animations
-```
-
----
-
-## 🔧 **CUSTOMIZATION**
-
-### **Adding New Stat Cards**
-```typescript
-// In page.tsx, add to statCards array
-{
-  title: 'New Metric',
-  value: stats.newMetric,
-  change: stats.newMetricGrowth,
-  changeLabel: 'vs last month',
-  icon: NewIcon,
-  color: 'blue',
-  onClick: () => handleNewMetricClick()
-}
-```
-
-### **Adding New Quick Actions**
-```typescript
-// In QuickActions.tsx, add to defaultActions
-{
-  id: 'new-action',
-  title: 'New Action',
-  description: 'Description of new action',
-  href: '/admin/new-page',
-  icon: NewIcon,
-  color: 'purple',
-  badge: 'New'
-}
-```
-
-### **Customizing Colors**
-```typescript
-// Update colorConfig in components
-const colorConfig = {
-  newColor: {
-    bg: 'bg-newcolor-50',
-    icon: 'text-newcolor-600',
-    border: 'border-l-newcolor-500',
-    gradient: 'from-newcolor-500 to-newcolor-600'
-  }
-};
-```
-
----
-
-## 🐛 **TROUBLESHOOTING**
-
-### **Common Issues**
-
-#### **1. Data Not Loading**
-```typescript
-// Check browser console for errors
-// Verify localStorage has adminToken
-// Check network tab for failed requests
-```
-
-#### **2. Responsive Issues**
-```css
-/* Check CSS media queries */
-/* Verify Tailwind breakpoints */
-/* Test on actual devices */
-```
-
-#### **3. Performance Issues**
-```typescript
-// Check React DevTools Profiler
-// Verify memo usage
-// Check for unnecessary re-renders
-```
-
-### **Debug Mode**
-```typescript
-// Add to localStorage for debug info
-localStorage.setItem('dashboard-debug', 'true');
-```
-
----
-
-## 🚀 **DEPLOYMENT**
-
-### **Production Checklist**
-- [ ] Build successful (`npm run build`)
-- [ ] No TypeScript errors
-- [ ] No ESLint warnings
-- [ ] Responsive tested on all breakpoints
-- [ ] Accessibility tested
-- [ ] Performance optimized
-- [ ] Error handling implemented
-- [ ] Loading states working
-
-### **Environment Variables**
-```bash
-# No additional env vars needed
-# Uses existing admin authentication
-```
-
----
-
-## 📈 **FUTURE ENHANCEMENTS**
-
-### **Phase 2 Features**
-- [ ] Real-time WebSocket updates
-- [ ] Advanced filtering and search
-- [ ] Customizable dashboard layouts
-- [ ] Export functionality
-- [ ] Dark mode support
-- [ ] Advanced charts and graphs
-
-### **Performance Improvements**
-- [ ] Virtual scrolling for large lists
-- [ ] Service worker for offline support
-- [ ] Progressive loading
-- [ ] Image optimization
-
----
-
-## 📞 **SUPPORT**
-
-### **Getting Help**
-1. Check this documentation
-2. Review component source code
-3. Test on different devices/browsers
-4. Check browser console for errors
-
-### **Reporting Issues**
-Include:
-- Browser and version
-- Screen size/device
-- Steps to reproduce
-- Expected vs actual behavior
-- Console errors (if any)
-
----
-
-## ✅ **TESTING CHECKLIST**
-
-### **Before Deployment**
-- [ ] All components render correctly
-- [ ] Responsive design works on all breakpoints
-- [ ] Accessibility features functional
-- [ ] Performance meets targets
-- [ ] Error handling works
-- [ ] Loading states display properly
-- [ ] Navigation links work
-- [ ] Data refreshes correctly
-- [ ] Animations smooth
-- [ ] No console errors
-
-### **Post Deployment**
-- [ ] Production build works
-- [ ] All features functional
-- [ ] Performance acceptable
-- [ ] No broken links
-- [ ] Analytics tracking (if applicable)
-
----
-
-**🎉 Dashboard is ready for production use!**
